+++ internal/dinosaur/pkg/wellknown/cloud_regions_test.go
@@ -1,4 +1,4 @@
-package clusters
+package wellknown
 
 import (
 	"testing"
