+++ internal/dinosaur/pkg/wellknown/doc.go
@@ -1,0 +1,2 @@
+// Package wellknown provides well-known constants and functions
+package wellknown
